% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lattice_ipr.R
\name{gg_lattice_ipr}
\alias{gg_lattice_ipr}
\title{Graphique de la série chronologique des IPR}
\usage{
gg_lattice_ipr(
  df,
  var_lattice,
  var_y,
  sup_500m = FALSE,
  metriques = FALSE,
  nb_colonnes = 3,
  max_axe_y = 8,
  inv_y = TRUE,
  annee_facteur = FALSE,
  titre_graphique = "",
  titre_y = "Métrique IPR",
  df_classes = classe_ipr,
  interactif = FALSE,
  largeur = 6,
  hauteur = 5,
  options = list()
)
}
\arguments{
\item{df}{Dataframe contenant les données IPR. Il doit contenir des variables "ipr"
et "annee" ainsi qu'une variable permettant d'identifier la station ou le point de
prélèvement.}

\item{var_lattice}{Nom de la variable servant à identifier les stations ou points.
Cette variable donnera les étiquettes du graphique.}

\item{var_y}{Nom de la variable contenant les valeurs IPR.}

\item{sup_500m}{Booléen. Indique si les stations sont situées à des altitudes inférieures
(sup_500m = FALSE, par défaut) ou bien supérieures à 500m car les seuils de classe
se qualité varient selon l'altitude. Comme ce paramètre est unique pour l'ensemble des
stations (ou points) représentées, il importe de s'assurer qu'elles sont toutes soit
au-dessus, soit en-dessous de 500m.}

\item{metriques}{Booléen. Si FALSE (défaut) les seuils de classes sont ceux de l'IPR.
Si TRUE, ils sont divisés par 7 pour s'adapter aux métriques.}

\item{nb_colonnes}{Entier. Nombre (maxi) de colonnes de graphiques s'il y a plusieurs stations.
Par défaut nb_colonnes = 6.}

\item{max_axe_y}{Numérique. Limite supérieure de l'axe des ordonnées. Par défaut max_axe_y = 40.}

\item{inv_y}{Booléen. Indique l'axe des ordonnées pointe vers le bas (TRUE, par défaut) ou
vers le haut.}

\item{annee_facteur}{Booléen. Indique si la variable annee doit être transformée en facteur.
Parfois utile pour la mise en forme de l'axe des abscisses (années).}

\item{titre_graphique}{Texte. Titre du graphique.}

\item{titre_y}{Texte. Titre de l'axe des ordonnées.}

\item{df_classes}{Dataframe contenant les limites de classes. Exemple : classe_ipr.}

\item{interactif}{Booléen. Indique si le graphique produit est statique (ggplot) ou interactif (produit avec le package ggiraph).}

\item{largeur, hauteur}{largeur et hauteur du graphique interactif (en pouces). Cela définit les dimensions relatives du graphique.}

\item{options}{liste d'options utilisée pour le rendu du graphique interactif (voir \code{\link[ggiraph]{girafe}})}

\item{station_sel}{Vecteur caractère indiquant les points ou stations à sélectionner.}
}
\value{
Un graphique statique ggplot2 ou un graphique interactif girafe.
}
\description{
La fonction utilise le référentiel "classe_ipr" de la base Aspe qui doit donc être
    chargé auparavant et complété par les codes couleurs avec ipr_completer_classes_couleur().
}
\examples{
\dontrun{
classe_ipr <- classe_ipr \%>\%
  ipr_completer_classes_couleur()

ipr_data \%>\%
  gg_temp_ipr(var_lattice = pop_libelle,
              var_y = ipr,
              station_sel = mes_pops,
              nb_colonnes = 4,
              max_axe_y = 60,
              inv_y = FALSE)
}
}
