% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lattice_line.R
\name{gg_lattice_line}
\alias{gg_lattice_line}
\title{Lattice diagram with ggplot with display options}
\usage{
gg_lattice_line(
  df,
  var_x,
  var_y,
  var_lattice,
  threshold_x = NULL,
  display_mean = TRUE,
  free_y = FALSE,
  lab_x,
  lab_y
)
}
\arguments{
\item{df}{Dataframe with the data. Must contain the variables described below.}

\item{var_x, var_y, var_lattice}{Variable for the x axis, y axis, and lattice wrapping.}

\item{threshold_x}{Numerical. If given, the line left from the threshold is dotted.
Typically used to distinguish time periods of distinct data reliability.}

\item{display_mean}{Boolean. Horizontal line for y mean value?}

\item{free_y}{Boolean. If TRUE, the y axis scale can vary across elementary plots.
If FALSE (default), the y axis scale is identical across elementary plots.}

\item{lab_x, lab_y}{String. Axis labels.}
}
\value{
The ggplot lattice diagram.
}
\description{
Lattice diagram with ggplot with display options
}
\examples{
\dontrun{
load(file = "tables_aspe.RData")
data <- classe_ipr \%>\%
  ipr_completer_classes_couleur()

ipr_data \%>\%
  gg_temp_ipr(var_id_sta = pop_libelle,
              var_ipr = ipr,
              station_sel = mes_pops,
              nb_colonnes = 4,
              max_axe_y = 60,
              inv_y = FALSE)
}
}
